Convert callers of vm_insert_pfn() to vm_fault_t return types
A lot of changes happened in the mm/memory.c area, like:
https://lore.kernel.org/linux-mm/20180828145728.11873-1-willy@infradead.org/

And for a long time already the *insert_pfn() API was moving toward a
vm_fault_t return type instead of using plain error codes to avoid
converting them on a per-driver basis into actual negative error
values.

Hence, the solution here is to treat VM_FAULT_NOPAGE as a regular
return code and do not error out in this case.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>