mali_kbase_mem_linux: fix link failure on mm_trace_rss_stat()
When building with Linux version >= 5.5 we get following error:
```
ERROR: modpost: "mm_trace_rss_stat" [mali_kbase.ko] undefined!
```
This is due to Linux commit:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=b3d1411b6726ea6930222f8f12587d89762477c6
that adds call to private mm_trace_rss_stat() in add_mm_counter(), but it's
not exported, so driver fails to link. Then when LINUX_VERSION_CODE >=
5.5.0 let's call only atomic_long_add() without mm_trace_rss_stat() that is
what basically add_mm_counter() does. mm_trace_rss_stat() is not exported
so we don't have a chance to make it link successfully unless we patch the
kernel, but this adds too much complexity and adds an external dependency
that we don't want. To do this let's define a local
mali_add_mm_counter() and call it in place of Linux add_mm_counter(),
this way everything gets easier to maintain and it's more readable.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>