mali_base_kernel.h: fix build failure on aarch64
Aarch64 PAGE_SHIFT is not 12 and this leads to build failure:
```
/home/giuliobenetti/br_reproduce/0e9207ab5806efc3885c27d1e98c7625574e57a3/output-1/build/mali-driver-3d697de9bce8bc68bf9408d5407873ce32d0f4fc/r8p0/drivers/gpu/arm/midgard/mali_kbase_mem_linux.c:2266:9: error: duplicate case value
 2266 |         case PFN_DOWN(BASEP_MEM_WRITE_ALLOC_PAGES_HANDLE):
      |         ^~~~
/home/giuliobenetti/br_reproduce/0e9207ab5806efc3885c27d1e98c7625574e57a3/output-1/build/mali-driver-3d697de9bce8bc68bf9408d5407873ce32d0f4fc/r8p0/drivers/gpu/arm/midgard/mali_kbase_mem_linux.c:2265:9: note: previously used here
 2265 |         case PFN_DOWN(BASEP_MEM_INVALID_HANDLE):
      |         ^~~~
```
This because PFN_DOWN(x) expands to ((x) >> PAGE_SHIFT), PAGE_SHIFT is not
always 12 and can be more like in Aarch64. This leads to override all
switch case PFN_DOWN(0) with a 0 and so compiler complains for duplicate
case values since all case values will be 0. Let's substitute magic number
12 with Linux PAGE_SHIFT macro to fix this.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>